+++ projects/games/battle/include/enemy.h
@@ -1,9 +1,8 @@
 #ifndef ENEMY_H
 #define ENEMY_H
 
-#include <stdint.h>
 #include "entity.h"
 
-extern entity enemy_create(uint16_t challenge_level);
+extern entity enemy_create(int challenge_level);
 
 #endif // ENEMY_H
